Changed texture to not be a template class, was stupid to make it one to begin with. Entities can now be both drawn and controlled. Let input begin

+++ Phoenix/PhoenixEngine.h
@@ -44,7 +44,7 @@ class Engine
     Engine(Modules *Module);
     ~Engine();
 
-    void Step(unsigned int _fps,bool _input[256], long mouse, long long _nFrameCount);
+    void Step(double _fps,bool _input[256], long mouse, long long _nFrameCount);
     void Render();
 };
